secure all end points

+++ src/controllers.php
@@ -97,6 +97,9 @@ $app->post('/todo/add', function (Request $request) use ($app) {
 
 
 $app->match('/todo/delete/{id}', function (Request $request, $id) use ($app) {
+    if (null === $user = $app['session']->get('user')) {
+        return $app->redirect('/login');
+    }
 
     $sql = "DELETE FROM todos WHERE id = '$id'";
     $app['db']->executeUpdate($sql);
@@ -111,6 +114,9 @@ $app->match('/todo/delete/{id}', function (Request $request, $id) use ($app) {
 
 
 $app->match('/todo/complete/{id}/{status}', function (Request $request, $id, $status) use ($app) { //change status based on what user has set
+    if (null === $user = $app['session']->get('user')) {
+        return $app->redirect('/login');
+    }
 
     $sql = "UPDATE todos SET completed = $status WHERE id = '$id'";
     $app['db']->executeUpdate($sql);
